Just run pyspelling with latest python

+++ app/install.sh
@@ -8,7 +8,10 @@ cd "${BASEDIR}"
 
 for PYVER in ${PYTHONVERS} ; do
   cd "${BASEDIR}/pip/${PYVER}"
-  "python${PYVER}" -m pip install -r requirements.txt
+  "python${PYVER}" -m pip install -r test_requirements.txt
+  if [ $(wc -l requirements.txt) -gt 0 ] ; then
+      "python${PYVER}" -m pip install -r requirements.txt
+  fi
   # Display installation
   "python${PYVER}" -m pip freeze
 done
